requirements.txt
easyocr
torch
httpx
tqdm
colorlog
pillow
cachetools